CNROM (mapper 3) support Fixed sprite jankiness due to evaluation timing

+++ src/Cartridge.h
@@ -20,7 +20,7 @@ public:
     std::ifstream& get_file() { return file; }
 
     uint32_t get_prg_size() const { return prg_size; }
-    uint16_t get_chr_size() const { return prg_size; }
+    uint32_t get_chr_size() const { return prg_size; }
     Memory* get_prg_rom() { return &prg_rom; }
     Memory* get_chr_rom() { return &chr_rom; }
     Memory* get_prg_ram() { return &prg_ram; }
@@ -38,7 +38,7 @@ private:
     std::ifstream file;
     uint16_t pos = 0;
     uint32_t prg_size;
-    uint16_t chr_size;
+    uint32_t chr_size;
     uint8_t mapper_num;
     bool flags[2][4];
